@@ -10,7 +10,8 @@
         <user-avatar user="Model.CurrentUser" class="mr-3 li-avatar rounded"></user-avatar>
 
         <div class="col-sm-6">
-            <input type="text" asp-for="@Model.NewCommentText" class="form-control" placeholder="Enter new comment..." />
+            <kendo-textbox for="@Model.NewCommentText" class="form-control" placeholder="Enter new comment...">
+            </kendo-textbox>
         </div>
         <button type="submit" class="btn btn-primary">Add</button>
     </div>
