@@ -4,7 +4,8 @@
     <input type="hidden" asp-for="@Model.ItemId" />
     <div class="form-row align-items-center">
         <div class="col-sm-6">
-            <input type="text" asp-for="@Model.NewTaskTitle" class="form-control" placeholder="Enter new task..." />
+            <kendo-textbox for="@Model.NewTaskTitle" class="form-control" placeholder="Enter new task...">
+            </kendo-textbox>
         </div>
         <button type="submit" class="btn btn-primary">Add</button>
     </div>
