@@ -6,7 +6,7 @@
     <title>@ViewData["Title"] - RPS</title>
 
     <link href="~/lib/bootstrap/dist/css/bootstrap.min.css" rel="stylesheet" />
-    <link href="~/lib/kendo/styles/kendo.default-v2.min.css" rel="stylesheet" />
+    <link href="~/lib/kendo/styles/default-main.css" rel="stylesheet" />
     <link href="~/css/site.css" rel="stylesheet" />
 
     @RenderSection("styles", required: false)
